chore: moves test files consts to the same file as rest of the logic
